requirements.txt
# virtio-serial RPC 依赖包
# Python >= 3.8

# 必需依赖
pyyaml>=6.0        # 配置文件解析

# 可选依赖（用于获取更详细的系统信息）
psutil>=5.9        # 系统信息获取
